% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_size.R
\name{cohenf2_to_r2}
\alias{cohenf2_to_r2}
\title{coenf2_to_r2}
\usage{
cohenf2_to_r2(cohensf2)
}
\arguments{
\item{cohensf2}{value of cohens f2}
}
\value{
corresponding r-squared value
}
\description{
Convert between Cohens f2 and r-squared
}
\examples{
cohenf2_to_r2(0.15)
}

