% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_size.R
\name{r2_to_cohenf2}
\alias{r2_to_cohenf2}
\title{r2_to_cohenf2}
\usage{
r2_to_cohenf2(r2)
}
\arguments{
\item{r2}{r-squared value}
}
\value{
corresponding Cohens f2
}
\description{
Convert between r-squared and Cohens f2
}
\examples{
r2_to_cohenf2(0.1304348)
}

