% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_size.R
\name{sample_size_2sided_2sample_t_test}
\alias{sample_size_2sided_2sample_t_test}
\title{sample_size_2sided_2sample_t_test}
\usage{
sample_size_2sided_2sample_t_test(cohensd, power, alpha)
}
\arguments{
\item{cohensd}{estimated Cohens-d}

\item{power}{desired statistical power}

\item{alpha}{desired alpha (p-value, alpha level, or type I error rate)}
}
\value{
required number of samples \strong{per group}
}
\description{
Calculate required sample size for a two sample two sided test
}
\examples{
sample_size_2sided_2sample_t_test(0.5,.8,0.05)
}

