% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_size.R
\name{sample_size_logistic_regression}
\alias{sample_size_logistic_regression}
\title{sample_size_logistic_regression}
\usage{
sample_size_logistic_regression(k, p)
}
\arguments{
\item{k}{the number of independent variables}

\item{p}{the percentage of samples for the lowest occuring class}
}
\value{
the number of samples needed
}
\description{
Conservative estimate of number of samples needed for logistic regression
}
\examples{
sample_size_logistic_regression(20,0.2)
}

