% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_size.R
\name{sample_size_multiple_regression}
\alias{sample_size_multiple_regression}
\title{sample_size_multiple_regression}
\usage{
sample_size_multiple_regression(nr_predictors, r2 = 0.2, effect_size = 0.15,
  power = 0.8, alpha = 0.05, verify = FALSE, ...)
}
\arguments{
\item{nr_predictors}{the number of independent covariates (i.e not including the target variable)}

\item{r2}{estimated r-squared, give either r2 or effect_size}

\item{effect_size}{effect sizes (also called Cohens f2) of 0.02, 0.15, and 0.35 are considered small, medium, and large, respectively}

\item{power}{the desired statistical power level}

\item{alpha}{p-value, alpha level, or type I error rate, usually 0.1, 0.05 or 0.01}

\item{verify}{should the sample size be 'verified' by simulation using MCMC, this can take a long time (on the order of 15 mins)}

\item{...}{Extra arguments to pass MBESS package}
}
\value{
the number of samples needed
}
\description{
Calculate the number of samples needed for mutiple linear regression
}
\details{
p-value, alpha level, or type I error rate. By convention, this value should be less than or equal to 0.05 to claim statistical significance
}
\examples{
sample_size_multiple_regression(20,effect_size=0.15,power=.8,alpha=0.05)
sample_size_multiple_regression(20,r2=0.1304348,power=.8,alpha=0.05)
}

